#proposal 86 tags and mobile (#3934)
* Implemented "Back-to-top"

Implementation of a "Back to top" icon in the bottom right corner on Desktop devices.  Currently only on Desktop due to the design / implementation challenge on mobile devices.

* Fixed bug #3931

Removed the "overflow: hidden" from the .App_content style in Apps.scss so that any overlays do not get cropped when opened.

Bug Fix #3931

* Carousel on Mobile Devices

Implemented the Community Carousel in Communities on Mobile Devices.

Users no long need to scroll past lots of pinned posts before seeing community content.

* Updated URL structure

Updated the link structure to improve SEO for community content.

Instead of 'hive-xxxxxx' in the URL, this is replaced by the communtiy name (switched to lower case and special characters removed/updated) or if this is invalid, the 1st tag of the user content.

This update also includes implementation of updated 'canonical-url' information so that Google prioritises the more SEO friendly URL.

* Fixed Bug #3932

Introduced a check on the 'settings' object (in Community Settings) so that it's not an empty object.  If no 'avatar_url' or 'cover_url' is set, the object will now reflect this.

* Fixed /Tags Page

Introduced a getTags request to retrieve Tags information from Hivemind in order to get the /tags page working again.

This is currently hard coded to retrieve (the maximum of) 250 tags but should be configured to accept this number as a variable to allow for the future implementation of a Tag Cloud or other functionality.

Community tags (i.e. 'hive-xxxxxx) are currently filtered out of the /Tags page but a future update could introduce these as their own section.

* Bug Fix #3935

Fix for bug #3935 so that the Preview clears the floating Markdown Style Guide.

* Removed Comments and Logs

Removed unnecessary comments and console.logs from 2 files.

* Removed Log

Removed console.log from PostList.jsx

* Mobile Bug Fix for Carousel

Fixed a bug on mobile devices where the 'next' and 'previous' arrows appeared unstyled when there is only 1 pinned post.

Fix: The arrows should not appear when there is only 1 pinned post.