#proposal 89 - images, Ukrainian and bug fixes (#3944)
* Voting Bar on Mobile Devices

Fix for Bug #3940

Current behaviour: The upvote bar appears off screen when a post / comment has received no votes.  When the upvote button is clicked, it is difficult to adjust the voting weight.   Once the vote has been submitted, the upvote and downvote buttons move to a more usable position.

This fix automatically puts the buttons in the more usable position permanently on mobile devices.

* Hide Resteems on Blogs

Fix for Bug #3936

Whilst the logic that decides whether the "Hide resteems" checkbox is used on a user's (Profile) Blog feed, the same logic is not used when filtering out posts.  This results in "Resteemed" posts being hidden from a user's blog if checked in their Friends Feed.

This fix adds the logic to check which page the user is on and doesn't hide resteems if they're not on a Friend's Feed.

* Missing Tags when Switching Communities

Fix for Bug #3928

The original code deliberately removed all of the existing tags in the draft so that a new 'community' tag (hive-xxxxxx) could be set.

This resulted in a confusing user experience where the Post Title and Body were populated, but the Post Tags were not.

This solution takes the existing tags from the draft, filters out community label (hive-xxxxxx) and then appends them to the community in the page.

As with the Title and Body, this retains the information after a page / state refresh.

* Terms of Service Wording

The existing wording:

"If you do not maintain a backup of your Account data outside of the Services, you will be may not be able to access STEEM, Steem Dollars and Steem Power associated with any Account maintained in your Account if we discontinue or deprecate the Services."

does not make sense so has been corrected to:

"If you do not maintain a backup of your Account data outside of the Services, you may not be able to access STEEM, Steem Dollars and Steem Power associated with your Account if we discontinue or deprecate the Services."

* Missing Tags when Switching Communities

Update to this bug fix to ensure that tags are not repeated / duplicated when editing a post.

* FAQs Update

Removed references to @cheetah, steemcleeaners and Utopian.io (which doesn't exist any more) from the FAQs.

* Fixed Tag Bug Introduced

A bug that was introduced when carrying Tags from one community to another.  It caused a JS error when editing a post that was not within a community.

* Implemented Ukrainian Language

Addition of Ukrainian translation and associated files.

* Images and New Tabs

Implementation of images automatically opening in a new tab as well as the ability for content creators to use target="_blank".

If a different target is set, it will be reset to blank for pages outside of steemit.com

* Ukrainian Language Fix

Fixed a couple of typos in the Ukrainian language file which prevented it from loading correctly.

* Ukrainian Language Updates

Updates as specified by o1eh in this comment:

https://steemit.com/ukraine-on-steem/@o1eh/stxkru

* Bug Fix

Fix for bug #3932

* Updated Image Opening

Improved the display of images so that instead of opening in a new window, they open in an overlay.

In addition to this, implemented the ability to zoom in and out of the image once loaded and reset it to its initial size.

It's also possible to grab the image to move the view around.

Also fixed a bug that opened post templates in a new window which isn't the desired behaviour.

* Community Settings

Implemented @only-dev-time fix to Community Settings

* Removed WIP

Removed code that was a work in progress.

* Fixed Image Link Bug

Fixed a bug introduced when implementing the overlay where images that already had a wrapper, tried to load the link URL as the image.  This has been fixed.

* Updated Primary Navigation

Moved Communities out of the tertiary navigation to make it more prominent.  This is particularly important for users that aren't logged in.

* Lightbox Upgrade

Complete overhaul of the Lightbox design for images.

* New Design
* Improved zooming in and out
* Mouse wheel zoom in and out
* Drag on mouse and touch events added

* Updated Zoom Calculations

There was a small error in the zoom calculations which adjusted the image ratio very slightly.  This has been corrected.

* Lightbox Fix

Adjusted the Lightbox height on mobile devices so that it was clear of the native device navigation controls.

* Translated Content

Updated lots of hard coded text and split them into their language translation files.

* Button Hover States

Implemented new colour styling and hover states on lightbox buttons - in both light mode and dark mode.

* Updated Voting Button Test

Since changing the "title" of the "Downvote" button to be localized, the test needed updating to reflect this.

* Image Display Bugs

Fixed a couple of bugs for edge case scenarios -

1. Relative URLs that wrap an image
2. External URL lightbox dimensions.

* Loading Improvements

Improved the image loading experience so that it doesn't flash and jerk across the screen.  Instead, there's a loading image which the photo grows out of.